+++ src/main/java/carracing/README.md
@@ -14,6 +14,7 @@
 
 - CarRacing 클래스 
   - 메인 기능 수행
+    - 주행을 시작하기 위해 질문을 호출한다.
     - 주행 자동차 목록 초기화한다.
     - 각 자동차 주행을 수행한다.
     - 우승자 자동차를 판단한다.
