+++ src/main/java/carracing/controller/CarRacing.java
@@ -1,11 +1,7 @@
 package carracing.controller;
 
-import carracing.domain.Car;
-import carracing.domain.Cars;
-import carracing.domain.RaceResult;
-import carracing.domain.Winners;
+import carracing.domain.*;
 import carracing.exception.CarNameFormatException;
-import carracing.util.RandomCondition;
 import carracing.view.InputInfo;
 import carracing.view.InputView;
 import carracing.view.ResultView;
