+++ src/test/java/carracing/CarRacingTest.java
@@ -11,20 +11,6 @@ import org.junit.jupiter.api.Test;
 public class CarRacingTest {
 
     @Test
-    @DisplayName("메인 기능 테스트")
-    void main() throws CarNameFormatException {
-
-        String carNames = "car1,car2,car3";
-        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분). : " + carNames);
-
-        String tryCount = "5";
-        System.out.println("시도할 회수는 몇 회 인가요? : " + tryCount);
-
-        CarRacing carRacing = new CarRacing();
-        carRacing.racingStart(new InputInfo(carNames, tryCount));
-    }
-
-    @Test
     @DisplayName("주어진 횟수 동안 경주 진행 여부 테스트")
     void tryCount() throws CarNameFormatException {
         String carNames = "car1,car2,car3";
