+++ src/test/java/carracing/PositionTest.java
@@ -1,6 +1,6 @@
 package carracing;
 
-import carracing.util.Position;
+import carracing.domain.Position;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
